refactor(app): decompose App.jsx
