+++ src/App.jsx
@@ -1,117 +1,56 @@
-import { Routes, Route, useNavigate, useLocation } from "react-router-dom";
-import {
-    Box,
-    Flex,
-    useDisclosure,
-    useColorMode,
-    IconButton,
-    useToast,
-} from "@chakra-ui/react";
+import { useState, useEffect, useCallback } from "react";
+import { useLocation } from "react-router-dom";
+import { useColorMode, useToast } from "@chakra-ui/react";
 
-import {
-    useState,
-    useEffect,
-    useCallback,
-    createContext,
-    useContext,
-} from "react";
-import { invoke } from "@tauri-apps/api/core";
 import { TemplateProvider } from "./utils/templates/templateContext";
 import { ApiToastProvider } from "./utils/helpers/apiToastContext";
-import Sidebar from "./components/sidebar/Sidebar";
-import LandingPage from "./pages/LandingPage";
-import PatientDetails from "./pages/PatientDetails";
-import Settings from "./pages/Settings";
-import Rag from "./pages/Rag";
-import ClinicSummary from "./pages/ClinicSummary";
-import OutstandingJobs from "./pages/OutstandingJobs";
+import { AppInitContext } from "./utils/context/appInit";
+import AppLayout from "./components/layout/AppLayout";
+import AppRoutes from "./components/layout/AppRoutes";
 import ConfirmLeaveModal from "./components/modals/ConfirmLeaveModal";
 import NewNoteModal from "./components/modals/NewNoteModal";
 import { handleError } from "./utils/helpers/errorHandlers";
-import {
-    handleLoadPatientDetails,
-    handleFetchPatientLetter,
-} from "./utils/patient/patientHandlers";
-import { usePatient } from "./utils/hooks/usePatient";
-import { useBreakpointValue } from "@chakra-ui/react";
-import theme from "./theme"; // Assuming theme is exported from here for ChakraProvider
-import SplashScreen from "./components/common/SplashScreen"; // Import SplashScreen
-import EncryptionSetup from "./components/setup/EncryptionSetup"; // Import EncryptionSetup
-import EncryptionUnlock from "./components/setup/EncryptionUnlock"; // Import EncryptionUnlock
-import ServerStartupLoader from "./components/setup/ServerStartupLoader"; // Import ServerStartupLoader
-import { settingsService } from "./utils/settings/settingsUtils"; // Import settingsService
-import { isTauri } from "./utils/helpers/apiConfig"; // Import isTauri
-import { encryptionApi } from "./utils/api/encryptionApi"; // Import encryptionApi
-
-// Context to track if app is initializing (server not ready yet)
-// This allows components to suppress error toasts during initialization
-export const AppInitContext = createContext({
-    isInitializing: false,
-});
-
-export const useAppInit = () => useContext(AppInitContext);
+import { handleLoadPatientDetails } from "./utils/patient/patientHandlers";
+import { usePatientSession } from "./utils/hooks/usePatientSession";
+import { useAppBootstrap } from "./utils/hooks/useAppBootstrap";
+import { useNavigationGuard } from "./utils/hooks/useNavigationGuard";
+import { useSidebarState } from "./utils/hooks/useSidebarState";
+import { useNewNoteFlow } from "./utils/hooks/useNewNoteFlow";
 
 function AppContent({ setIsInitializing }) {
-    const [showSplashScreen, setShowSplashScreen] = useState(undefined);
-    const [isLoadingSplashCheck, setIsLoadingSplashCheck] = useState(true);
-    const [refreshKey, setRefreshKey] = useState(0);
-    const [newNoteKey, setNewNoteKey] = useState(0);
     const [isModified, setIsModified] = useState(false);
-
-    // Encryption state
-    const [, setEncryptionStatus] = useState(null);
-    const [showEncryptionSetup, setShowEncryptionSetup] = useState(false);
-    const [showEncryptionUnlock, setShowEncryptionUnlock] = useState(false);
-    const [, setIsLoadingEncryptionCheck] = useState(true);
-    const [showServerStartupLoader, setShowServerStartupLoader] =
-        useState(false);
-    const [isInGracePeriod, setIsInGracePeriod] = useState(true);
-
-    // App initialization state - true when server is not ready yet
-    // This is used to suppress error toasts during initialization
-    // isInGracePeriod provides a buffer after server is ready to suppress
-    // toasts from initial API calls that may fail before app stabilizes
-    const isInitializing =
-        showEncryptionSetup ||
-        showEncryptionUnlock ||
-        showServerStartupLoader ||
-        isInGracePeriod;
-
-    // Update the parent state whenever isInitializing changes
-    useEffect(() => {
-        if (setIsInitializing) {
-            setIsInitializing(isInitializing);
-        }
-    }, [isInitializing, setIsInitializing]);
-
-    const { isOpen, onOpen, onClose } = useDisclosure();
-    const {
-        isOpen: isNewNoteOpen,
-        onOpen: onOpenNewNote,
-        onClose: onCloseNewNote,
-    } = useDisclosure();
-    const [pendingNavigation, setPendingNavigation] = useState(null);
-    const navigate = useNavigate();
-    const toast = useToast();
-    const location = useLocation();
-    const { colorMode, toggleColorMode } = useColorMode();
+    const [refreshKey, setRefreshKey] = useState(0);
     const [isFromOutstandingJobs, setIsFromOutstandingJobs] = useState(false);
-    const [resetLetter, setResetLetter] = useState(null);
 
-    // Use the patient hook for all patient-related state
+    // App-level patient "session": briefcase patient + shared selectedDate +
+    // new-note actions. The editor (PatientDetails) owns its own copy.
     const {
         patient,
         setPatient,
         selectedDate,
         setSelectedDate,
-        finalCorrespondence,
-        setFinalCorrespondence,
         createNewPatient,
         findPatients,
         loadSelectedPatient,
-        templateKey,
-        setTemplateKey,
-    } = usePatient();
+    } = usePatientSession();
+
+    const bootstrap = useAppBootstrap();
+    const nav = useNavigationGuard(isModified, setIsModified);
+    const newNote = useNewNoteFlow({
+        createNewPatient,
+        guardedNavigate: nav.guardedNavigate,
+    });
+    const { isSidebarCollapsed, toggleSidebar, isSmallScreen } =
+        useSidebarState();
+    const { colorMode, toggleColorMode } = useColorMode();
+    const toast = useToast();
+    const location = useLocation();
+
+    useEffect(() => {
+        if (setIsInitializing) {
+            setIsInitializing(bootstrap.isInitializing);
+        }
+    }, [bootstrap.isInitializing, setIsInitializing]);
 
     const fetchPatientDetailsWrapper = useCallback(
         async (noteId) => {
@@ -136,516 +75,83 @@ function AppContent({ setIsInitializing }) {
         }
     }, [location, fetchPatientDetailsWrapper]);
 
-    const startNewNote = useCallback(async () => {
-        await createNewPatient();
-        setNewNoteKey((k) => k + 1);
-        if (resetLetter) {
-            resetLetter();
-        }
-    }, [createNewPatient, resetLetter]);
-
-    const handleNewPatient = () => {
-        toast.closeAll();
-        onOpenNewNote();
-    };
-
-    const completeNewNote = ({ cameFromSearch } = {}) => {
-        setNewNoteKey((k) => k + 1);
-        if (resetLetter) {
-            resetLetter();
-        }
-        onCloseNewNote();
-        handleNavigation("/new-note", {
-            viaModal: true,
-            cameFromSearch: Boolean(cameFromSearch),
-        });
-    };
+    const refreshSidebar = useCallback(() => {
+        setRefreshKey((prev) => prev + 1);
+    }, []);
 
     const handleSelectPatient = (
         selectedPatient,
         fromOutstandingJobs = false,
     ) => {
         setIsFromOutstandingJobs(fromOutstandingJobs);
-        if (isModified) {
-            setPendingNavigation({ path: `/note/${selectedPatient.id}` });
-            onOpen();
-        } else {
-            navigate(`/note/${selectedPatient.id}`);
-        }
+        nav.guardedNavigate(`/note/${selectedPatient.id}`);
     };
 
-    const CollapseIcon = ({ boxSize = "20px" }) => (
-        <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-            strokeWidth="1.5"
-            width={boxSize}
-            height={boxSize}
-        >
-            <rect
-                x="3"
-                y="3"
-                width="18"
-                height="18"
-                rx="5"
-                ry="5"
-                stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                fill="none"
-            />
-            <path
-                d="M9.5 21V3"
-                stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-            />
-        </svg>
-    );
-
-    const handleNavigation = (path, state) => {
-        toast.closeAll();
-        if (isModified) {
-            setPendingNavigation({ path, state });
-            onOpen();
-        } else {
-            setIsModified(false);
-            navigate(path, state ? { state } : undefined);
-        }
-    };
-
-    const confirmNavigation = () => {
-        onClose();
-        if (pendingNavigation) {
-            const { path, state } = pendingNavigation;
-            setIsModified(false);
-            navigate(path, state ? { state } : undefined);
-            setPendingNavigation(null);
-        }
-    };
-
-    const cancelNavigation = () => {
-        onClose();
-        setPendingNavigation(null);
-    };
-
-    const refreshSidebar = useCallback(() => {
-        setRefreshKey((prev) => prev + 1);
-    }, []);
-
-    useEffect(() => {
-        const handleBeforeUnload = (e) => {
-            if (isModified) {
-                e.preventDefault();
-                e.returnValue =
-                    "You have unsaved changes. Are you sure you want to leave?";
-            }
-        };
-
-        window.addEventListener("beforeunload", handleBeforeUnload);
-        return () =>
-            window.removeEventListener("beforeunload", handleBeforeUnload);
-    }, [isModified]);
-
-    // Sidebar Items
-    const toggleSidebar = useCallback(() => {
-        setIsSidebarCollapsed((prev) => !prev);
-    }, []);
-
-    // This will return true for "base" and "sm" breakpoints, false for larger screens
-    const defaultCollapsed = useBreakpointValue({
-        base: true,
-        sm: true,
-        md: false,
-    });
-
-    // Initialize sidebar state with the breakpoint-dependent value
-    const [isSidebarCollapsed, setIsSidebarCollapsed] =
-        useState(defaultCollapsed);
-
-    const isSmallScreen = useBreakpointValue({ base: true, md: false });
-
-    // Update sidebar state whenever the breakpoint changes
-    useEffect(() => {
-        if (defaultCollapsed !== undefined) {
-            setIsSidebarCollapsed(defaultCollapsed);
-        }
-    }, [defaultCollapsed]);
-
-    // Splash screen logic
-    const checkSplashStatus = useCallback(async (options = {}) => {
-        const { maxRetries = 5, retryDelay = 500 } = options;
-        let lastError;
-        let success = false;
-
-        for (let attempt = 0; attempt < maxRetries; attempt++) {
-            try {
-                await settingsService.fetchUserSettings((userData) => {
-                    if (
-                        userData &&
-                        typeof userData.has_completed_splash_screen ===
-                            "boolean"
-                    ) {
-                        setShowSplashScreen(
-                            !userData.has_completed_splash_screen,
-                        );
-                    } else {
-                        setShowSplashScreen(true); // Default to showing splash if flag is missing/invalid
-                    }
-                });
-                // Success - exit the retry loop
-                success = true;
-                break;
-            } catch (error) {
-                console.warn(
-                    `Error checking splash screen status (attempt ${attempt + 1}/${maxRetries}):`,
-                    error,
-                );
-                lastError = error;
-                // Wait before retrying (except on the last attempt)
-                if (attempt < maxRetries - 1) {
-                    await new Promise((resolve) =>
-                        setTimeout(resolve, retryDelay),
-                    );
-                }
-            }
-        }
-
-        if (!success) {
-            // All retries failed
-            console.error(
-                "Error checking splash screen status after retries:",
-                lastError,
-            );
-            setShowSplashScreen(true); // Default to showing splash on error
-        }
-
-        setIsLoadingSplashCheck(false);
-    }, []);
-
-    // Only check splash status on mount if NOT in Tauri
-    // In Tauri, we wait until after encryption unlock is complete
-    useEffect(() => {
-        if (!isTauri()) {
-            checkSplashStatus();
-        } else {
-            // In Tauri, mark splash check as complete immediately
-            // We'll check it after encryption unlock
-            setIsLoadingSplashCheck(false);
-        }
-    }, [checkSplashStatus]);
-
-    // In non-Tauri environments, disable grace period immediately since server is already running
-    useEffect(() => {
-        if (!isTauri()) {
-            console.log("Non-Tauri environment detected");
-            setIsInGracePeriod(false);
-        }
-    }, []);
-
-    const handleSplashComplete = () => {
-        setShowSplashScreen(false);
-    };
-
-    // Encryption check logic - only for Tauri builds
-    useEffect(() => {
-        if (!isTauri()) {
-            setIsLoadingEncryptionCheck(false);
-            return;
-        }
-
-        const checkEncryptionStatus = async () => {
-            try {
-                const status = await encryptionApi.getStatus();
-                setEncryptionStatus(status);
-
-                // Determine what to show based on status
-                if (!status.has_setup && !status.has_database) {
-                    // First-time setup - no encryption, no database
-                    setShowEncryptionSetup(true);
-                } else if (status.has_setup && !status.has_keychain) {
-                    // Has encrypted data but not unlocked
-                    // Start the server in warm mode (it will wait for passphrase)
-                    try {
-                        await invoke("start_server_command");
-                        console.log(
-                            "Server started in warm mode, waiting for passphrase",
-                        );
-                    } catch (e) {
-                        console.warn("Failed to warm start server:", e);
-                    }
-                    setShowEncryptionUnlock(true);
-                }
-                // If has_keychain is true, proceed to app normally
-            } catch (error) {
-                console.error("Error checking encryption status:", error);
-                // On error, allow proceeding (may be in dev mode)
-            } finally {
-                setIsLoadingEncryptionCheck(false);
-            }
-        };
-
-        checkEncryptionStatus();
-    }, []);
-
-    const handleEncryptionSetupComplete = () => {
-        setShowEncryptionSetup(false);
-        // Show server startup loader while server initializes
-        setShowServerStartupLoader(true);
-    };
-
-    const handleEncryptionUnlockComplete = () => {
-        setShowEncryptionUnlock(false);
-        // Show server startup loader while server initializes
-        setShowServerStartupLoader(true);
-    };
-
-    const handleServerReady = () => {
-        setShowServerStartupLoader(false);
-        // Server is ready, now check splash screen status
-        checkSplashStatus({ maxRetries: 3, retryDelay: 300 });
-        // Keep suppressing toasts for a short grace period after server is ready
-        // This allows initial API calls to fail gracefully without showing error toasts
-        setTimeout(() => {
-            setIsInGracePeriod(false);
-        }, 2000); // 2 second grace period
-    };
-
-    const handleServerError = (error) => {
-        console.error("Server startup error:", error);
-        setShowServerStartupLoader(false);
-        // Show error toast
-        toast({
-            title: "Server Error",
-            description: error.message || "Failed to start the server",
-            status: "error",
-            duration: 5000,
-            isClosable: true,
-        });
-        // Go back to unlock screen
-        setShowEncryptionUnlock(true);
-    };
-
-    if (isLoadingSplashCheck) {
-        return null; // or a loading spinner
+    if (bootstrap.gate) {
+        return bootstrap.gate;
     }
 
-    // Show encryption setup first (before splash screen)
-    if (showEncryptionSetup) {
-        return <EncryptionSetup onComplete={handleEncryptionSetupComplete} />;
-    }
-
-    // Show encryption unlock if needed
-    if (showEncryptionUnlock) {
-        return <EncryptionUnlock onComplete={handleEncryptionUnlockComplete} />;
-    }
-
-    // Show server startup loader while server initializes
-    if (showServerStartupLoader) {
-        return (
-            <ServerStartupLoader
-                onReady={handleServerReady}
-                onError={handleServerError}
-            />
-        );
-    }
-
-    if (showSplashScreen) {
-        return <SplashScreen onComplete={handleSplashComplete} />;
-    }
     return (
-        <Flex position="relative">
-            {/* Floating hamburger button for small screens */}
-            {isSmallScreen && isSidebarCollapsed && (
-                <IconButton
-                    icon={<CollapseIcon />}
-                    onClick={toggleSidebar}
-                    position="fixed"
-                    top="6"
-                    left="6"
-                    zIndex="101"
-                    aria-label="Toggle sidebar"
-                    className="dark-toggle"
-                />
-            )}
-
-            <Sidebar
-                onNewPatient={handleNewPatient}
-                onSelectPatient={handleSelectPatient}
-                selectedDate={selectedDate}
-                setSelectedDate={setSelectedDate}
-                refreshKey={refreshKey}
-                handleNavigation={handleNavigation}
-                isCollapsed={isSidebarCollapsed}
-                toggleSidebar={toggleSidebar}
+        <>
+            <AppLayout
                 isSmallScreen={isSmallScreen}
+                isCollapsed={isSidebarCollapsed}
                 colorMode={colorMode}
-                toggleColorMode={toggleColorMode}
-            />
-
-            <Box
-                flex="1"
-                ml={isSmallScreen ? "0" : isSidebarCollapsed ? "96px" : "236px"}
-                minH="100vh"
-                transition="margin-left 0.3s ease"
-                bg={
-                    isTauri()
-                        ? colorMode === "light"
-                            ? "#232634"
-                            : "#1e2030"
-                        : "transparent"
-                }
-                display="flex"
-                flexDirection="column"
+                toggleSidebar={toggleSidebar}
+                sidebarProps={{
+                    onNewPatient: newNote.openNewNoteModal,
+                    onSelectPatient: handleSelectPatient,
+                    selectedDate,
+                    setSelectedDate,
+                    refreshKey,
+                    handleNavigation: nav.guardedNavigate,
+                    isCollapsed: isSidebarCollapsed,
+                    toggleSidebar,
+                    isSmallScreen,
+                    colorMode,
+                    toggleColorMode,
+                }}
             >
-                {/* Tauri titlebar drag region for macOS */}
-                {isTauri() && (
-                    <Box
-                        data-tauri-drag-region
-                        height="25px"
-                        position="fixed"
-                        top="0"
-                        right="0"
-                        left={
-                            isSmallScreen
-                                ? "0"
-                                : isSidebarCollapsed
-                                  ? "96px"
-                                  : "236px"
-                        }
-                        zIndex="1000"
-                        transition="left 0.3s ease"
-                    />
-                )}
-
-                <Box
-                    m="0px"
-                    borderRadius="0px"
-                    p={isTauri() ? "6" : "0"}
-                    pt={
-                        isSmallScreen && isTauri()
-                            ? "50px"
-                            : isTauri()
-                              ? "6"
-                              : "0"
-                    }
-                    className="main-bg"
-                    height="100vh"
-                    overflowY="auto"
-                    position="relative"
-                >
-                    <Routes>
-                        <Route
-                            path="/new-note"
-                            element={
-                                <PatientDetails
-                                    key={`new-note-${newNoteKey}`}
-                                    patient={patient}
-                                    setPatient={setPatient}
-                                    selectedDate={selectedDate}
-                                    refreshSidebar={refreshSidebar}
-                                    setIsModified={setIsModified}
-                                    finalCorrespondence={finalCorrespondence}
-                                    setFinalCorrespondence={
-                                        setFinalCorrespondence
-                                    }
-                                    templateKey={templateKey}
-                                    setTemplateKey={setTemplateKey}
-                                    onResetLetter={setResetLetter}
-                                    onStartNewNote={startNewNote}
-                                />
-                            }
-                        />
-                        <Route
-                            path="/note/:id"
-                            element={
-                                <PatientDetails
-                                    patient={patient}
-                                    setPatient={setPatient}
-                                    selectedDate={selectedDate}
-                                    refreshSidebar={refreshSidebar}
-                                    setIsModified={setIsModified}
-                                    finalCorrespondence={finalCorrespondence}
-                                    setFinalCorrespondence={
-                                        setFinalCorrespondence
-                                    }
-                                    templateKey={templateKey}
-                                    setTemplateKey={setTemplateKey}
-                                    onStartNewNote={startNewNote}
-                                />
-                            }
-                        />
-                        <Route path="/" element={<LandingPage />} />
-                        <Route path="/settings" element={<Settings />} />
-                        <Route path="/rag" element={<Rag />} />
-                        <Route
-                            path="/clinic-summary"
-                            element={
-                                <ClinicSummary
-                                    selectedDate={selectedDate}
-                                    handleSelectPatient={handleSelectPatient}
-                                    refreshSidebar={refreshSidebar}
-                                />
-                            }
-                        />
-                        <Route
-                            path="/outstanding-jobs"
-                            element={
-                                <OutstandingJobs
-                                    handleSelectPatient={(patient) =>
-                                        handleSelectPatient(patient, true)
-                                    }
-                                    refreshSidebar={refreshSidebar}
-                                />
-                            }
-                        />
-                    </Routes>
-                </Box>
-            </Box>
+                <AppRoutes
+                    patient={patient}
+                    setPatient={setPatient}
+                    selectedDate={selectedDate}
+                    refreshSidebar={refreshSidebar}
+                    setIsModified={setIsModified}
+                    onResetLetter={newNote.setResetLetter}
+                    onStartNewNote={newNote.startNewNote}
+                    newNoteKey={newNote.newNoteKey}
+                    handleSelectPatient={handleSelectPatient}
+                />
+            </AppLayout>
             <NewNoteModal
-                isOpen={isNewNoteOpen}
-                onClose={onCloseNewNote}
+                isOpen={newNote.isNewNoteOpen}
+                onClose={newNote.closeNewNoteModal}
                 patient={patient}
                 setPatient={setPatient}
                 createNewPatient={createNewPatient}
                 findPatients={findPatients}
                 loadSelectedPatient={loadSelectedPatient}
                 selectedDate={selectedDate}
-                onComplete={completeNewNote}
+                onComplete={newNote.completeNewNote}
             />
             <ConfirmLeaveModal
-                isOpen={isOpen}
-                onClose={cancelNavigation}
-                confirmNavigation={confirmNavigation}
+                isOpen={nav.isLeaveOpen}
+                onClose={nav.cancelNavigation}
+                confirmNavigation={nav.confirmNavigation}
             />
-        </Flex>
+        </>
     );
 }
 
-// Wrap the entire app with TemplateProvider and AppInitContext
 function App() {
-    return (
-        <AppInitContext.Provider value={{ isInitializing: false }}>
-            <AppWithInitState />
-        </AppInitContext.Provider>
-    );
-}
-
-// Inner component that provides the actual isInitializing state
-function AppWithInitState() {
-    // Get the initialization state from AppContent
-    // We need to restructure slightly to pass this down
-    const [isInitializingState, setIsInitializingState] = useState(true);
+    const [isInitializing, setIsInitializing] = useState(true);
 
     return (
-        <AppInitContext.Provider
-            value={{ isInitializing: isInitializingState }}
-        >
+        <AppInitContext.Provider value={{ isInitializing }}>
             <ApiToastProvider>
                 <TemplateProvider>
-                    <AppContent setIsInitializing={setIsInitializingState} />
+                    <AppContent setIsInitializing={setIsInitializing} />
                 </TemplateProvider>
             </ApiToastProvider>
         </AppInitContext.Provider>
