+++ src/pages/LandingPage.jsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import { Box } from "@chakra-ui/react";
 import DashboardChat from "../components/dashboard/DashboardChat";
 import DisclaimerModal from "../components/modals/DisclaimerModal";
-import { useAppInit } from "../App";
+import { useAppInit } from "../utils/context/appInit";
 
 const LandingPage = () => {
   const { isInitializing } = useAppInit();
